Port message sending to `ZulipClient`
