+++ src/zulip.rs
@@ -1,4 +1,4 @@
-mod api;
+pub mod api;
 pub mod client;
 
 use crate::db::notifications::add_metadata;
@@ -11,6 +11,7 @@ use crate::handlers::project_goals::{self, ping_project_goals_owners};
 use crate::handlers::Context;
 use crate::team_data::{people, teams};
 use crate::utils::pluralize;
+use crate::zulip::api::{MessageApiResponse, Recipient};
 use crate::zulip::client::ZulipClient;
 use anyhow::{format_err, Context as _};
 use rust_team_data::v1::TeamKind;
@@ -265,7 +266,7 @@ fn handle_command<'a>(
                                 };
 
                                 if project_goals::check_project_goal_acl(&ctx.github, gh_id).await? {
-                                    ping_project_goals_owners(&ctx.github, false, threshold, &format!("on {next_update}"))
+                                    ping_project_goals_owners(&ctx.github, &ctx.zulip, false, threshold, &format!("on {next_update}"))
                                         .await
                                         .map_err(|e| format_err!("Failed to await at this time: {e:?}"))?;
                                     return Ok(None);
@@ -275,7 +276,7 @@ fn handle_command<'a>(
                                     ));
                                 }
                             }
-                            Some("docs-update") => return trigger_docs_update(message_data),
+                            Some("docs-update") => return trigger_docs_update(message_data, &ctx.zulip),
                             _ => {}
                         }
                     }
@@ -687,7 +688,7 @@ async fn execute_for_other_user(
         },
         content: &message,
     }
-    .send(ctx.github.raw())
+    .send(&ctx.zulip)
     .await;
 
     if let Err(err) = res {
@@ -697,87 +698,10 @@ async fn execute_for_other_user(
     Ok(Some(output))
 }
 
-#[derive(Copy, Clone, serde::Serialize)]
-#[serde(tag = "type")]
-#[serde(rename_all = "snake_case")]
-pub enum Recipient<'a> {
-    Stream {
-        #[serde(rename = "to")]
-        id: u64,
-        topic: &'a str,
-    },
-    Private {
-        #[serde(skip)]
-        id: u64,
-        #[serde(rename = "to")]
-        email: &'a str,
-    },
-}
-
-impl Recipient<'_> {
-    pub fn narrow(&self) -> String {
-        match self {
-            Recipient::Stream { id, topic } => {
-                // See
-                // https://github.com/zulip/zulip/blob/46247623fc279/zerver/lib/url_encoding.py#L9
-                // ALWAYS_SAFE without `.` from
-                // https://github.com/python/cpython/blob/113e2b0a07c/Lib/urllib/parse.py#L772-L775
-                //
-                // ALWAYS_SAFE doesn't contain `.` because Zulip actually encodes them to be able
-                // to use `.` instead of `%` in the encoded strings
-                const ALWAYS_SAFE: &str =
-                    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789_-~";
-
-                let mut encoded_topic = String::new();
-                for ch in topic.bytes() {
-                    if !(ALWAYS_SAFE.contains(ch as char)) {
-                        write!(encoded_topic, ".{:02X}", ch).unwrap();
-                    } else {
-                        encoded_topic.push(ch as char);
-                    }
-                }
-                format!("stream/{}-xxx/topic/{}", id, encoded_topic)
-            }
-            Recipient::Private { id, .. } => format!("pm-with/{}-xxx", id),
-        }
-    }
-
-    pub fn url(&self) -> String {
-        format!("{}/#narrow/{}", *ZULIP_URL, self.narrow())
-    }
-}
-
-#[cfg(test)]
-fn check_encode(topic: &str, expected: &str) {
-    const PREFIX: &str = "stream/0-xxx/topic/";
-    let computed = Recipient::Stream { id: 0, topic }.narrow();
-    assert_eq!(&computed[..PREFIX.len()], PREFIX);
-    assert_eq!(&computed[PREFIX.len()..], expected);
-}
-
-#[test]
-fn test_encode() {
-    check_encode("some text with spaces", "some.20text.20with.20spaces");
-    check_encode(
-        " !\"#$%&'()*+,-./",
-        ".20.21.22.23.24.25.26.27.28.29.2A.2B.2C-.2E.2F",
-    );
-    check_encode("0123456789:;<=>?", "0123456789.3A.3B.3C.3D.3E.3F");
-    check_encode(
-        "@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_",
-        ".40ABCDEFGHIJKLMNOPQRSTUVWXYZ.5B.5C.5D.5E_",
-    );
-    check_encode(
-        "`abcdefghijklmnopqrstuvwxyz{|}~",
-        ".60abcdefghijklmnopqrstuvwxyz.7B.7C.7D~.7F",
-    );
-    check_encode("áé…", ".C3.A1.C3.A9.E2.80.A6");
-}
-
 #[derive(serde::Serialize)]
-pub struct MessageApiRequest<'a> {
-    pub recipient: Recipient<'a>,
-    pub content: &'a str,
+pub(crate) struct MessageApiRequest<'a> {
+    pub(crate) recipient: Recipient<'a>,
+    pub(crate) content: &'a str,
 }
 
 impl<'a> MessageApiRequest<'a> {
@@ -785,65 +709,9 @@ impl<'a> MessageApiRequest<'a> {
         self.recipient.url()
     }
 
-    pub async fn send(&self, client: &reqwest::Client) -> anyhow::Result<MessageApiResponse> {
-        let bot_api_token = env::var("ZULIP_API_TOKEN").expect("ZULIP_API_TOKEN");
-
-        #[derive(serde::Serialize)]
-        struct SerializedApi<'a> {
-            #[serde(rename = "type")]
-            type_: &'static str,
-            to: String,
-            #[serde(skip_serializing_if = "Option::is_none")]
-            topic: Option<&'a str>,
-            content: &'a str,
-        }
-
-        let resp = client
-            .post(format!("{}/api/v1/messages", *ZULIP_URL))
-            .basic_auth(&*ZULIP_BOT_EMAIL, Some(&bot_api_token))
-            .form(&SerializedApi {
-                type_: match self.recipient {
-                    Recipient::Stream { .. } => "stream",
-                    Recipient::Private { .. } => "private",
-                },
-                to: match self.recipient {
-                    Recipient::Stream { id, .. } => id.to_string(),
-                    Recipient::Private { email, .. } => email.to_string(),
-                },
-                topic: match self.recipient {
-                    Recipient::Stream { topic, .. } => Some(topic),
-                    Recipient::Private { .. } => None,
-                },
-                content: self.content,
-            })
-            .send()
-            .await
-            .context("fail sending Zulip message")?;
-
-        let status = resp.status();
-
-        if !status.is_success() {
-            let body = resp
-                .text()
-                .await
-                .context("fail receiving Zulip API response (when sending a message)")?;
-
-            anyhow::bail!(body)
-        }
-
-        let resp: MessageApiResponse = resp
-            .json()
-            .await
-            .context("fail receiving the JSON Zulip Api reponse (when sending a message)")?;
-
-        Ok(resp)
+    pub(crate) async fn send(&self, client: &ZulipClient) -> anyhow::Result<MessageApiResponse> {
+        client.send_message(self.recipient, self.content).await
     }
-}
-
-#[derive(Debug, serde::Deserialize)]
-pub struct MessageApiResponse {
-    #[serde(rename = "id")]
-    pub message_id: u64,
 }
 
 #[derive(Debug)]
@@ -1144,7 +1012,7 @@ async fn post_waiter(
         },
         content: waiting.primary,
     }
-    .send(ctx.github.raw())
+    .send(&ctx.zulip)
     .await?;
 
     for reaction in waiting.emoji {
@@ -1160,10 +1028,11 @@ async fn post_waiter(
     Ok(None)
 }
 
-fn trigger_docs_update(message: &Message) -> anyhow::Result<Option<String>> {
+fn trigger_docs_update(message: &Message, zulip: &ZulipClient) -> anyhow::Result<Option<String>> {
     let message = message.clone();
     // The default Zulip timeout of 10 seconds can be too short, so process in
     // the background.
+    let zulip = zulip.clone();
     tokio::task::spawn(async move {
         let response = match docs_update().await {
             Ok(None) => "No updates found.".to_string(),
@@ -1179,7 +1048,7 @@ fn trigger_docs_update(message: &Message) -> anyhow::Result<Option<String>> {
             recipient,
             content: &response,
         };
-        if let Err(e) = message.send(&reqwest::Client::new()).await {
+        if let Err(e) = message.send(&zulip).await {
             log::error!("failed to send Zulip response: {e:?}\nresponse was:\n{response}");
         }
     });
